refatorando código e corrigindo problemas de merge em searchClient
