+++ .gitignore
@@ -2,4 +2,4 @@
 .idea/workspace.xml
 .DS_Store
 *.class
-.target
+target/
